refactor: move word into it's own directory
