+++ src/lib/data_types/word/word.c
