+++ src/lib/word_bank.h
@@ -1,7 +1,7 @@
 #ifndef WORD_BANK_H
 #define WORD_BANK_H
 
-#include "./data_types/word.h"
+#include "./data_types/word/word.h"
 
 void print_something(void);
