+++ src/main.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include "lib/data_types/word.h"
 #include "lib/word_bank.h"
 
 int main(int argc, char ** argv) {
